+++ include/employee_cab_search.h
@@ -9,11 +9,11 @@ using namespace std;
 class EmployeeCabSearch {
 private:
     vector<Cab>& cabs;
-    static double calculate_distance(pair<int, int> location1, pair<int, int> location2);
+    static double calculate_distance(pair<double, double> location1, pair<double, double> location2);
 
 public:
     EmployeeCabSearch(vector<Cab>& cabs);
-    vector<Cab> suggest_nearby_cabs(pair<int, int> employee_location);
+    vector<Cab> suggest_nearby_cabs(pair<double, double> employee_location);
 };
 
 #endif // EMPLOYEE_CAB_SEARCH_H
